Update meetingsGuide.md
There is some inconsistency in this document regarding capitalization particularly with « proposer ». Many of these terms should not be capitalized, but if they are, there needs to be clarity and consistency across all documentation. In general, such haphazard capitalization makes reading and understanding a document more difficult. Advice requested. I can make changes as needed.