Update projectVerification.md
Is there a Giveth « Projects Page »? Above was stated that approved projects will be on the homepage (shouldn’t be capital but I left it for now) and on « the Project’s Page ». If that is referring to a Giveth page, it should not have an apostrophe. If it is referring to the project’s page itself (of the project), it should not be capitalized. Thanks!