Update givstream.md
Thé « and » is confusing at the end. If the person doesn’t need to fulfill BOTH « vote » and « provide liquidity », then it should be « or ». The person can « donate » or « vote » or « provide liquidity ». Correct? If it is either 1 or 2 & 3, then the « and » stays. If it is 1 or 2 or 3, then you can take out the « and » and the comma. Otherwise insert a comma after the « donate » line and include « or » on both « donate » and « vote » lines. Hope that makes sense! See below. :)

The GIVstream nurtures the GIVeconomy at inception by having only 10% of the total supply of GIV liquid and transferrable to start. As the GIVeconomy grows & stabilizes, more GIV becomes liquid and available for everyone. We want to empower those who support the Giveth ecosystem with steadily increasing governance rights, this includes participants who:
- [Donate to Verified Projects](https://giveth.io/projects), or
- Vote in the [GIVgarden](https://giv.giveth.io/givgarden), or
- Provide liquidity in the [GIVfarm](https://giv.giveth.io/givfarm)

If so, it should be:

The GIVstream nurtures the GIVeconomy at inception by having only 10% of the total supply of GIV liquid and transferrable to start. As the GIVeconomy grows & stabilizes, more GIV becomes liquid and available for everyone. We want to empower those who support the Giveth ecosystem with steadily increasing governance rights, this includes participants who:
- [Donate to Verified Projects](https://giveth.io/projects), and
- Vote in the [GIVgarden](https://giv.giveth.io/givgarden), and
- Provide liquidity in the [GIVfarm](https://giv.giveth.io/givfarm)